Replace local components
