--- src/components/Header/Header.vue
@@ -1,19 +1,0 @@
-<template>
-	<header class="Header">
-		<h1 class="Header__brand-name">
-			Vue Masterclass
-		</h1>
-	</header>
-</template>
-
-<style>
-.Header {
-    border-bottom: 2px solid black;
-}
-.Header__brand-name {
-    font-size: 1.4rem;
-    line-height: 1;
-    margin: 0;
-    padding: 1rem;
-}
-</style>
