+++ src/views/Home.vue
@@ -1,10 +1,23 @@
 <script setup lang="ts">
-import HelloWorld from '../components/HelloWorld.vue'
-import Header from '../components/Header/Header.vue'
+import { useAuthStore } from '@/stores/auth.store'
+import { useCounterStore } from '@/stores/counter.store'
+import { Button, Header } from 'vue-masterclass-library'
+
+const auth = useAuthStore()
+const counter = useCounterStore()
+
+const handleDecrement = () => {
+	counter.decrement()
+}
+
+const handleIncrement = () => {
+	counter.increment()
+}
 </script>
 
 <template>
-	<Header />
+	<Header brand-name="Vue Masterclass" />
+
 	<div class="wrapper">
 		<div>
 			<a
@@ -24,11 +37,70 @@ import Header from '../components/Header/Header.vue'
 					alt="Vue logo">
 			</a>
 		</div>
-		<HelloWorld msg="Vite + Vue" />
+		<h1>Vite + Vue</h1>
+
+		<div class="card">
+			<p>Count is {{ counter.count }}</p>
+
+			<!-- Decrement button -->
+			<Button
+				variant="secondary"
+				@click="handleDecrement">
+				decrement
+			</Button>
+
+			<!-- Increment button -->
+			<Button
+				variant="primary"
+				@click="handleIncrement">
+				increment
+			</Button>
+
+			<p>
+				Edit
+				<code>components/HelloWorld.vue</code> to test HMR
+			</p>
+		</div>
+
+		<div class="card">
+			<p v-if="auth.user">
+				User: {{ auth.userFullName }}
+			</p>
+			<Button v-if="!auth.user" @click="auth.login">
+				Login
+			</Button>
+			<Button v-if="auth.user" @click="auth.logout">
+				Logout
+			</Button>
+		</div>
+
+		<p>
+			Check out
+			<a
+				href="https://vuejs.org/guide/quick-start.html#local"
+				target="_blank">
+				create-vue
+			</a>, the official Vue + Vite starter
+		</p>
+		<p>
+			Install
+			<a
+				href="https://github.com/johnsoncodehk/volar"
+				target="_blank">
+				Volar
+			</a>
+			in your IDE for a better DX
+		</p>
+		<p class="read-the-docs">
+			Click on the Vite and Vue logos to learn more
+		</p>
 	</div>
 </template>
 
 <style scoped>
+button + button {
+	margin-left: .5rem;
+}
 .logo {
   height: 6em;
   padding: 1.5em;
@@ -39,6 +111,9 @@ import Header from '../components/Header/Header.vue'
 }
 .logo.vue:hover {
   filter: drop-shadow(0 0 2em #42b883aa);
+}
+.read-the-docs {
+	color: #888;
 }
 .wrapper {
   margin: 0 auto;
